requirements.txt
opencv-python
pillow
face_recognition
mediapipe
numpy
tensorflow
faker
python-escpos